Make function signatures consistent between inheriting classes.
Overriding classes must provide methods with similar signatures.
Functions must either be static in both or none of the classes,
and functions must have the same amount of parameters in both classes.

The DBObject object has a create function that is never used,
and can't be matched to all children.  It's better to simply remove it.